+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased](https://github.com/XaviArnaus/pyxavi/)
+## [v0.6.0](https://github.com/XaviArnaus/pyxavi/releases/tag/v0.6.0) - 2023-10-27
 
 - Added a `storage.get_keys_in()`.
 - Added a `Url` class for operations over URLs. At this point, just a shorthand for URL cleaning.